bench: add gst runner
